requirements.txt
fastapi~=0.116.1
uvicorn[standard]
gymnasium
numpy~=1.23.5
pydantic~=2.11.7
gym~=0.26.2
websockets~=15.0.1
sqlalchemy~=2.0.41
pydantic-settings~=2.10.1
python-dotenv~=1.1.1